requirements.txt
PyOpenGL==3.1.9
PyQT5==5.15.11
PyWavefront==1.3.3